+++ src/actions/LocationActions.js
@@ -7,17 +7,7 @@ class LocationActions {
   }
 
   fetchLocations() {
-    // we dispatch an event here so we can have "loading" state.
     this.dispatch();
-
-    LocationsFetcher.fetch()
-      .then((locations) => {
-        // we can access other actions within our action through `this.actions`
-        this.actions.updateLocations(locations);
-      })
-      .catch((errorMessage) => {
-        this.actions.locationsFailed(errorMessage);
-      });
   }
 
   locationsFailed(errorMessage) {
