+++ src/stores/LocationStore.js
@@ -1,5 +1,6 @@
 var alt = require('../alt');
 var LocationActions = require('../actions/LocationActions');
+var LocationSource = require('../sources/LocationSource');
 var FavoritesStore = require('./FavoritesStore');
 
 class LocationStore {
@@ -15,7 +16,9 @@ class LocationStore {
 
     this.exportPublicMethods({
       getLocation: this.getLocation
-    })
+    });
+
+    this.exportAsync(LocationSource);
   }
 
   handleUpdateLocations(locations) {
@@ -24,8 +27,6 @@ class LocationStore {
   }
 
   handleFetchLocations() {
-    // reset the array while we're fetching new locations so React can
-    // be smart and render a spinner for us since the data is empty.
     this.locations = [];
   }
